test(claude): split the compat file by topic behind one fake API
claude_thinking_compat_test.dart had grown from 1427 to 3989 lines over
#943 and this branch: one group, 72 tests, six unrelated topics, four
near-identical "bind a server, reply as Claude, proxy the official host"
helpers and twelve more servers hand-rolled inside tests.

The scaffolding is now one helper, `captureClaudeExchange`, in
test/support/claude_test_api.dart, with the config and block builders the
files share. The tests move verbatim - same 72 names, same assertions -
into files named for what they check:

- claude_thinking_compat_test.dart      thinking, effort, sampling, caching
- claude_hosted_tools_request_test.dart what each endpoint is sent as tools
- claude_history_replay_test.dart       hand-offs, pause_turn, images, replay
- claude_container_turn_test.dart       containers, stored turns, downloads

Five tests keep their own server: they need binary bodies, error statuses,
path dispatch, a dropped connection or a slow reply.
